forum: remove timeline extraction and display
