+++ next-app/app/forum/[forumId]/[postId]/page.jsx
@@ -342,125 +342,6 @@ export default function Page({ params, searchParams }) {
                                         flexGrow: 1,
                                     }}
                                 >
-                                    {protocolForumPost?.timeline && (
-                                        <>
-                                            <h1
-                                                style={{
-                                                    fontSize: "16px",
-                                                    fontWeight: 600,
-                                                    marginBottom: "20px",
-                                                }}
-                                            >
-                                                Timeline
-                                            </h1>
-                                            <div
-                                                style={{
-                                                    width: "100%",
-                                                    display: "flex",
-                                                    flexDirection: "row",
-                                                    justifyContent:
-                                                        "flex-start",
-                                                    alignItems: "center",
-                                                    marginBottom: "20px",
-                                                }}
-                                            >
-                                                <div
-                                                    style={{
-                                                        width: "10px",
-                                                        height: "10px",
-                                                        borderRadius: "50%",
-                                                        backgroundColor:
-                                                            protocolForumPost
-                                                                ?.timeline
-                                                                .primaryColor,
-                                                    }}
-                                                />
-                                                <div
-                                                    style={{
-                                                        height: "2px",
-                                                        width: `calc(${protocolForumPost?.timeline?.percentProgress}% - 10px)`,
-                                                        backgroundColor:
-                                                            protocolForumPost
-                                                                ?.timeline
-                                                                ?.primaryColor,
-                                                    }}
-                                                />
-                                                <div
-                                                    style={{
-                                                        width: "10px",
-                                                        height: "10px",
-                                                        borderRadius: "50%",
-                                                        backgroundColor: "#fff",
-                                                        boxShadow:
-                                                            "0px 4px 21.8px 0px rgba(0,0,0,0.75)",
-                                                    }}
-                                                />
-                                                <div
-                                                    style={{
-                                                        width: `calc(${100 - protocolForumPost?.timeline?.percentProgress}% - 20px)`,
-                                                        height: "2px",
-                                                        border: "none",
-                                                        borderTop: `1px dashed ${protocolForumPost?.timeline?.secondaryColor}`,
-                                                    }}
-                                                />{" "}
-                                                <div
-                                                    style={{
-                                                        width: "10px",
-                                                        height: "10px",
-                                                        borderRadius: "50%",
-                                                        backgroundColor:
-                                                            protocolForumPost
-                                                                ?.timeline
-                                                                .secondaryColor,
-                                                    }}
-                                                />
-                                            </div>
-                                            <div
-                                                style={{
-                                                    display: "flex",
-                                                    flexDirection: "row",
-                                                    justifyContent:
-                                                        "flex-start",
-                                                    alignItems: "center",
-                                                    marginTop: "5px",
-                                                }}
-                                            >
-                                                <p
-                                                    style={{
-                                                        fontSize: "12px",
-                                                        color: "#A2A2AE",
-                                                    }}
-                                                >
-                                                    {
-                                                        protocolForumPost
-                                                            ?.timeline?.start
-                                                    }
-                                                </p>
-                                                <p
-                                                    style={{
-                                                        fontSize: "12px",
-                                                        color: "#A2A2AE",
-                                                        marginLeft: `calc(${protocolForumPost?.timeline?.percentProgress}% - 30px)`,
-                                                    }}
-                                                >
-                                                    Today
-                                                </p>
-                                                <p
-                                                    style={{
-                                                        fontSize: "12px",
-                                                        color: "#A2A2AE",
-                                                        marginLeft: `calc(${100 - protocolForumPost?.timeline?.percentProgress}% - 35px)`,
-                                                    }}
-                                                >
-                                                    {
-                                                        protocolForumPost
-                                                            ?.timeline?.end
-                                                    }
-                                                </p>
-                                            </div>
-                                        </>
-                                    )}
-
                                     <h1
                                         style={{
                                             fontSize: "16px",
